requirements.txt
schedule==0.6.0
cmcrameri==1.2
click==7.1.2
pydeck==0.6.1
pandas==1.0.5
matplotlib==3.2.2
numpy==1.18.5
EQTransformer==0.1.59
obspy==1.2.2
pandas_read_xml==0.0.9
plotly==4.14.3
streamlit==0.78.0
tensorflow==2.0.0
mysql-connector-python